requirements.txt
streamlit
opencv-python-headless
pytesseract
ultralytics
pyttsx3
Pillow
numpy
